pyproject.toml
[project]
name = "ebsagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "fastapi>=0.117.1",
    "fastmcp>=2.12.2",
    "oci>=2.160.0",
    "oracledb>=3.3.0",
    "python-dotenv>=1.1.1",
]
